several changes and fixes. Read the ChangeLog
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@694 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/insets/insettext.C
@@ -1019,7 +1019,8 @@ void InsetText::setPos(Painter & pain, int x, int y) const
     cursor.y = top_baseline;
     y += cursor.y;
     for(unsigned int i = 1;
-	((cursor.y + rows[i - 1].desc) < y) && (i < rows.size() - 1); ++i) {
+	(long(cursor.y + rows[i - 1].desc) < y)
+		&& (i < rows.size() - 1); ++i) {
 	cursor.y = rows[i].baseline;
 	cursor.pos = rows[i].pos;
 	actrow = i;
